Create FlowPanel.Simulate class
Added nested class Simulate in FlowPanel to handle simulation. This is
the class that will be threaded now, not FlowPanel, mostly so FlowPanel has
more controller duties and Flow has less.

Also added play button in Flow. At the moment each click of play button results
in one simulation timestep only.

Water.flow() was adjusted to allow depth to be set to zero.