fix for thread objects, improve simulation functionality
Added a Thread[] field and NUM_THREADS constant to FlowPanel to handle
threads. There was a Simulate field instead which ofc led to
NullPointerExceptions when i attempted to use it as the thread.

Added boolean cancelled field to FlowPanel and cancel check in while
loop to Simulate.run(). Now the threads can actually be stopped.

Play button now has different function depending on if simulation has
been paused or reset. isReset boolean field added to Flow.