Fix Simulate call to getPermute()
Simulate.run() was calling Grid.getPermute() with the index of the map
instead of index of the permute list, throwing outOfBounds exception.
Changed a for loop to go from 0 to length of list, using new
Grid.subLen() method, instead of lo to hi. Simulate no longer needs lo
and hi, just thread number.