Add water conservation testing
Testing that water no water is created or removed when it should be
helps identify race conditions.

Add Water variables for added, removed, and current count. In all
methods that add/remove, modify the vars accordingly, and have function
to count up all the water.

In FlowPanel, get and print the values at each pause